Update validate-direct-form.js
Fixes a validation error on the checkout page that caused problems when customers chose a different payment method and the crypto field was still required. This validation error also caused problems for the checkout page when the plugin was disabled, making it impossible to click the 'place order' button. (Also credits to Thinkbeat Solutions)